requirements.txt
streamlit
chromadb
langchain
langchain-community
langchainhub
python-dotenv
pypdf
jq


